% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovr.R
\name{recovr}
\alias{recovr}
\title{Recover RStudio Files}
\usage{
recovr(project_path, out_folder = NULL, force = FALSE)
}
\arguments{
\item{project_path}{The project directory from which to recover files.
Typically a path on disk to the directory. Use the special value
\code{NULL} to recover files from sessions not associated with projects,
and \code{"all"} to recover files from every recently used project and
session.}

\item{out_folder}{The path to which recovered files should be written.
Defaults to the folder "recovered_XXX" in the root of the project
when run on a single project (where XXX is the current date/time),
or the current working directory if not.}

\item{force}{Whether \code{out_folder} should be overwritten if it already exists.}
}
\value{
A data frame listing the recovered files.
}
\description{
Recovers unsaved changes or abandoned auto-save copies of files
previously opened in RStudio.
}
\examples{
\dontrun{
# Recover files in the current project to the temporary directory
recovr()

# Recover all recent projects/files to a custom folder
recovr(project_path = "all", out_folder = "~/scratch/rstudio")

}

}
